Javadoc, move into main package
